requirements.txt
streamlit
torch
torchvision
transformers
opencv-python
numpy
Pillow
supervision
timm
shutil
